% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/racmo.R
\docType{package}
\name{racmo-package}
\alias{racmo}
\alias{racmo-package}
\title{racmo: Download and Process RACMO NetCDF Data}
\description{
The \strong{racmo} package provides tools to download and process meteorological data
from RACMOv2.3 (KNMI Data Platform) used for STOWA drought statistics.
It includes functions to:
\itemize{
\item Retrieve file lists from the KNMI API: \link{get_filenames}
\item Download and convert NetCDF files into projected SpatRaster stacks: \link{download_and_load_raster}
}
}
\details{
Data source: \link{url} (see \url{https://dataplatform.knmi.nl/dataset/knmi23-droogtestatistiek-1-0})
}
\section{Functions}{

\describe{
\item{\link{get_filenames}}{Retrieve filenames from the KNMI API.}
\item{\link{download_and_load_raster}}{Download and process NetCDF files into raster stacks.}
}
}

\section{Constants}{

\describe{
\item{\link{url}}{URL constant pointing to KNMI dataset.}
}
}

\author{
\strong{Maintainer}: Kees van Immerzeel \email{c.h.van.immerzeel@gmail.com} (\href{https://orcid.org/0000-0003-0926-3248}{ORCID})

}
\keyword{internal}
